+++ lib/ruby-clock.rb
@@ -1,6 +1,7 @@
 require "ruby-clock/version"
 require "ruby-clock/rake"
 require "ruby-clock/shell"
+require "ruby-clock/around_actions"
 require 'rufus-scheduler'
 require 'singleton'
 
@@ -8,23 +9,12 @@ class RubyClock
   include Singleton
   include RubyClock::Rake
   include RubyClock::Shell
+  include RubyClock::AroundActions
 
-  attr_accessor :on_error, :around_actions
+  attr_accessor :on_error
 
   def initialize
-    @around_actions = []
-
-    def schedule.around_trigger(job_info, &job_proc)
-      RubyClock.instance.call_with_around_action_stack(
-        RubyClock.instance.around_actions.reverse,
-        job_proc,
-        job_info
-      )
-    end
-  end
-
-  def freeze_around_actions
-    @around_actions.freeze
+    set_up_around_actions
   end
 
   def wait_seconds
@@ -61,18 +51,4 @@ class RubyClock
     schedule.resume
     schedule.join
   end
-
-  def call_with_around_action_stack(wrappers, job_proc, job_info)
-    case wrappers.count
-    when 0
-      job_proc.call(job_info)
-    else
-      call_with_around_action_stack(
-        wrappers[1..],
-        Proc.new{ wrappers.first.call(job_proc, job_info) },
-        job_info
-      )
-    end
-  end
-
 end
